--- a/crates/subspace-proof-of-space/src/chiapos/table/types.rs
+++ b/crates/subspace-proof-of-space/src/chiapos/table/types.rs
@@ -6,6 +6,7 @@ use core::mem;
 use core::ops::Range;
 use derive_more::{Add, AddAssign, From, Into};
 
+/// Stores data in lower bits
 #[derive(Debug, Copy, Clone, Ord, PartialOrd, Eq, PartialEq, From, Into, Add, AddAssign)]
 #[repr(transparent)]
 pub(in super::super) struct X(u32);
@@ -49,6 +50,7 @@ impl X {
     }
 }
 
+/// Stores data in lower bits
 #[derive(Debug, Default, Copy, Clone, Ord, PartialOrd, Eq, PartialEq, From, Into)]
 #[repr(transparent)]
 pub(in super::super) struct Y(u32);
@@ -102,6 +104,7 @@ impl Position {
     pub(in super::super) const ONE: Self = Self(1);
 }
 
+/// Stores data in lower bits
 #[derive(Debug, Copy, Clone, Ord, PartialOrd, Eq, PartialEq)]
 #[repr(transparent)]
 pub(in super::super) struct Metadata<const K: u8, const TABLE_NUMBER: u8>(
@@ -137,7 +140,7 @@ where
     EvaluatableUsize<{ metadata_size_bytes(K, TABLE_NUMBER) }>: Sized,
 {
     /// If used incorrectly, will truncate information, it is up to implementation to ensure `u128`
-    /// only contains data in least significant bits and fits into internal byte array of `Metadata`
+    /// only contains data in lower bits and fits into internal byte array of `Metadata`
     fn from(value: u128) -> Self {
         Self(
             value.to_be_bytes()[mem::size_of::<u128>() - metadata_size_bytes(K, TABLE_NUMBER)..]