Small improvements as a follow-up to https://github.com/subspace/subspace/pull/1677

These are really tiny, but they are measurable.

k17:
```
chia/table/parallel     time:   [21.585 ms 21.744 ms 21.867 ms]
```
k22:
```
chia/table/single       time:   [4.6323 s 4.6357 s 4.6393 s]
chia/table/parallel     time:   [732.06 ms 756.15 ms 779.80 ms]
```

So far struggling to get it any lower. Tried all of the options I could think of for now. If I find something else I'll open a separate PR.

I have an idea for improvements with a slightly different API that I'll try next. A lot of time is apparently spent on allocation/deallocation during benchmarks, we can improve things by reusing caches across calls, but for that we need a stateful API.

### Code contributor checklist:
* [x] I have read, understood and followed [contributing guide](https://github.com/subspace/subspace/blob/main/CONTRIBUTING.md)